eth: mcux: Do not set carrier ON if interface is not known
It is possible that the network interface is not yet initialized
when status of the PHY changes. In this case we must not call
net_eth_carrier_on() as that will cause a crash.
This was noticed with mimxrt1050_evk board.

Fixes: #21257

(cherry picked from commit e124c1cd34938f02bbe879533c6bd29b5ff53707)

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>