drivers: eth: mcux: Remove useless SMIRead which conflicts with SMIWrite
Starting a SMI write operation without waiting for completion of the
preceding SMI read operation cause the write operation to fail if
the time between the 2 operations it too short. This leads to the
PHY being in an unusable state on the MIMRT1060-EVK eval board.

To fix the problem the SMI read operation has been removed since
the read value is not used.

Signed-off-by: Armand Ciejak <armandciejak@users.noreply.github.com>