drivers: flash: flexspi_nor: Statically allocate the bounce buffer
This avoids problems when using the driver in mcuboot.
Since the buffer is never freed it is not necessary to
allocate it with malloc, avoiding a possible error
in init function.
Use a u32_t pointer instead of a void pointer to have
the same type as the one used by the FlexSPI function,
It also guarantees the alignment of the address.

Signed-off-by: Armand Ciejak <armand@riedonetworks.com>