drivers: flash: flexspi_nor: Add missing software reset
According to i.MX RT1060 manual in chapter
"27.5.11.1 Command Abort and Suspend" it is required to issue
a software reset after erase/program commands.

This is present in NXP SDK, but was forgotten when the driver was
ported to Zephyr.

This allows having the sequence in proper order in
critical_section_leave function.

Signed-off-by: Armand Ciejak <armand@riedonetworks.com>